refactor(networking): extract AsyncPolicyBase ABC (step 2 of #109) (#113)
* refactor(networking): extract AsyncPolicyBase ABC (step 2 of #109)

All shared async policy logic (circuit breaking, retry loop, backoff,
rate limiting, proxy pool, metadata assembly) moved into AsyncPolicyBase.

AsyncHttpClient and AsyncCurlHttpClient become thin subclasses — each
implements: aclose, _is_transport_exception, _is_retryable_exception,
_execute_attempt, _handle_request_exception.

Key design decisions vs the sync pair:
- _execute_attempt replaces _apply_proxy in the base: httpx creates a
  fresh client per proxy and closes it in finally; curl mutates
  _session.proxies. Both patterns fit cleanly into one abstract method.
- _build_meta remains concrete in the base (curl-compatible defaults);
  AsyncHttpClient overrides it for httpx field names (reason_phrase,
  str(url), redirected flag).
- No _proxies abstract property needed: proxy handling is fully
  encapsulated in _execute_attempt, so the base never touches _session.

Closes step 2 of #109.


* docs(async_policy_base): restore end-to-start stamp semantics note

_enforce_rate_limit had an explanatory docstring in the original
AsyncHttpClient describing that it only waits and that all timestamp
stamping is done in the _request finally block. Lost during extraction;
restored on the base class where it now applies to both transports.


---------
